ARM: dts: msm: Add a vote for CPU to GPU Config path for SM8150
We should have a vote for CPU to GPU config path
when we are trying to access any GPU config registers.
This will make sure that all the clocks in the CPU to
GPU config paths are ON and we are not riding on
someone's else vote for this path.

Change-Id: I1964570a36a5d1045def161eb5f4f5fbecda7b95
Signed-off-by: Prakash Kamliya <pkamliya@codeaurora.org>
(cherry picked from commit 1eae8a60d40902cb39d8d0a6ac614b32ac4ccb95)